fix some warming

+++ 05-ScenceManager/CMushroom.cpp
@@ -5,8 +5,8 @@
 bool CMushroom::isMoney = false;
 bool CMushroom::isStart = false;
 bool CMushroom::isRun = false;
-int CMushroom::xBox = 0;
-int CMushroom::yBox = 0;
+float CMushroom::xBox = 0;
+float CMushroom::yBox = 0;
 CMushroom * CMushroom::__instance = NULL;
 CMushroom::CMushroom() : CGameObject()
 {
@@ -120,7 +120,7 @@ void CMushroom::Update(DWORD dt, vector<LPGAMEOBJECT> *coObjects)
 			if (y >= 306)
 			{
 				//vy = 0.01;
-				vx = -0.055;
+				vx = -0.055f;
 				y =306;
 				//x -= dx;
 			}
@@ -128,14 +128,14 @@ void CMushroom::Update(DWORD dt, vector<LPGAMEOBJECT> *coObjects)
 			{
 				//x = 0;
 				vx = 0;
-				y += 0.5*dt;
+				y += 0.5f*dt;
 			}
 			else
 			{
 				y = 188;
 				//y += 0.005*dt;
 				//x -= 0.4*vx;
-				vx = -0.04;
+				vx = -0.04f;
 			}
 		}
 		x += dx;
